+++ src/LinqSql.Tests/LinqSql.Tests.csproj
@@ -68,6 +68,18 @@
     <Compile Include="Expressions\Implementations\WhereExpressionTests.cs" />
     <Compile Include="Expressions\Implementations\TableExpressionTests.cs" />
     <Compile Include="Expressions\MockExpressionVisitor.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableSingleTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableSingleOrDefaultTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableFirstTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableFirstOrDefaultTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableLastTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableLastOrDefaultTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableContainsTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableJoinTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableSkipTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableAggregateTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableOrderByTests.cs" />
+    <Compile Include="Queryable\SqlQueryableLinqTests\SqlQueryableWhereTests.cs" />
     <Compile Include="Visitors\QueryVisitor\QueryTests.cs" />
     <Compile Include="Visitors\QueryVisitor\SqlQueryContextTests.cs" />
     <Compile Include="Visitors\QueryVisitor\SqlQueryVisitorTests.cs" />
