+++ src/LinqSql.Tests/Queryable/SqlQueryableLinqTests/SqlQueryableJoinTests.cs
@@ -1,0 +1,156 @@
+﻿using System.Data.Common;
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace System.Linq.Sql.Tests
+{
+    using Sqlite;
+
+    [TestClass]
+    public class SqlQueryableJoinTests
+    {
+        private readonly DbConnection connection = ConnectionTestHelper.CreatePopulatedConnection();
+
+        [TestMethod]
+        public void SqlQueryable_Join_SelectBoth()
+        {
+            // Prepare the test data
+            IQueryable<Record> outer = new SqliteQueryable(connection, "Course", new[] { "Id", "Name" });
+            IQueryable<Record> inner = new SqliteQueryable(connection, "CourseStudent", new[] { "Id", "CourseId", "StudentId" });
+
+            // Perform the test operation
+            Record[] records = outer
+                .Join(inner, x => x["Course"]["Id"], x => x["CourseStudent"]["CourseId"], (o, i) => o | i)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(8, records.Length);
+            foreach (Record record in records)
+            {
+                Assert.AreEqual(2, record.Count);
+                Assert.IsTrue(record["Course"].ContainsKey("Id"));
+                Assert.IsTrue(record["Course"].ContainsKey("Name"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("Id"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("CourseId"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("StudentId"));
+                Assert.AreEqual(record["Course"]["Id"], record["CourseStudent"]["CourseId"]);
+            }
+        }
+
+        [TestMethod]
+        public void SqlQueryable_Join_SelectOuter()
+        {
+            // Prepare the test data
+            IQueryable<Record> outer = new SqliteQueryable(connection, "Course", new[] { "Id", "Name" });
+            IQueryable<Record> inner = new SqliteQueryable(connection, "CourseStudent", new[] { "Id", "CourseId", "StudentId" });
+
+            // Perform the test operation
+            Record[] records = outer
+                .Join(inner, x => x["Course"]["Id"], x => x["CourseStudent"]["CourseId"], (o, i) => o)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(8, records.Length);
+            foreach (Record record in records)
+            {
+                Assert.AreEqual(1, record.Count);
+                Assert.IsTrue(record["Course"].ContainsKey("Id"));
+                Assert.IsTrue(record["Course"].ContainsKey("Name"));
+            }
+        }
+
+        [TestMethod]
+        public void SqlQueryable_Join_SelectInner()
+        {
+            // Prepare the test data
+            IQueryable<Record> outer = new SqliteQueryable(connection, "Course", new[] { "Id", "Name" });
+            IQueryable<Record> inner = new SqliteQueryable(connection, "CourseStudent", new[] { "Id", "CourseId", "StudentId" });
+
+            // Perform the test operation
+            Record[] records = outer
+                .Join(inner, x => x["Course"]["Id"], x => x["CourseStudent"]["CourseId"], (o, i) => i)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(8, records.Length);
+            foreach (Record record in records)
+            {
+                Assert.AreEqual(1, record.Count);
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("Id"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("CourseId"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("StudentId"));
+            }
+        }
+
+        [TestMethod]
+        public void SqlQueryable_Join_Predicate_SelectBoth()
+        {
+            // Prepare the test data
+            IQueryable<Record> outer = new SqliteQueryable(connection, "Course", new[] { "Id", "Name" });
+            IQueryable<Record> inner = new SqliteQueryable(connection, "CourseStudent", new[] { "Id", "CourseId", "StudentId" });
+
+            // Perform the test operation
+            Record[] records = outer
+                .Join(inner, (o, i) => o["Course"]["Id"] == i["CourseStudent"]["CourseId"], (o, i) => o | i)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(8, records.Length);
+            foreach (Record record in records)
+            {
+                Assert.AreEqual(2, record.Count);
+                Assert.IsTrue(record["Course"].ContainsKey("Id"));
+                Assert.IsTrue(record["Course"].ContainsKey("Name"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("Id"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("CourseId"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("StudentId"));
+                Assert.AreEqual(record["Course"]["Id"], record["CourseStudent"]["CourseId"]);
+            }
+        }
+
+        [TestMethod]
+        public void SqlQueryable_Join_Predicate_SelectOuter()
+        {
+            // Prepare the test data
+            IQueryable<Record> outer = new SqliteQueryable(connection, "Course", new[] { "Id", "Name" });
+            IQueryable<Record> inner = new SqliteQueryable(connection, "CourseStudent", new[] { "Id", "CourseId", "StudentId" });
+
+            // Perform the test operation
+            Record[] records = outer
+                .Join(inner, (o, i) => o["Course"]["Id"] == i["CourseStudent"]["CourseId"], (o, i) => o)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(8, records.Length);
+            foreach (Record record in records)
+            {
+                Assert.AreEqual(1, record.Count);
+                Assert.IsTrue(record["Course"].ContainsKey("Id"));
+                Assert.IsTrue(record["Course"].ContainsKey("Name"));
+            }
+        }
+
+        [TestMethod]
+        public void SqlQueryable_Join_Predicate_SelectInner()
+        {
+            // Prepare the test data
+            IQueryable<Record> outer = new SqliteQueryable(connection, "Course", new[] { "Id", "Name" });
+            IQueryable<Record> inner = new SqliteQueryable(connection, "CourseStudent", new[] { "Id", "CourseId", "StudentId" });
+
+            // Perform the test operation
+            Record[] records = outer
+                .Join(inner, (o, i) => o["Course"]["Id"] == i["CourseStudent"]["CourseId"], (o, i) => i)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(8, records.Length);
+            foreach (Record record in records)
+            {
+                Assert.AreEqual(1, record.Count);
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("Id"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("CourseId"));
+                Assert.IsTrue(record["CourseStudent"].ContainsKey("StudentId"));
+            }
+        }
+    }
+}
