+++ src/LinqSql.Tests/Queryable/SqlQueryableLinqTests/SqlQueryableSingleOrDefaultTests.cs
@@ -1,0 +1,40 @@
+﻿using System.Data.Common;
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace System.Linq.Sql.Tests
+{
+    using Sqlite;
+
+    [TestClass]
+    public class SqlQueryableSingleOrDefaultTests
+    {
+        private readonly DbConnection connection = ConnectionTestHelper.CreatePopulatedConnection();
+
+        [TestMethod]
+        [ExpectedException(typeof(InvalidOperationException))]
+        public void SqlQueryable_SingleOrDefault()
+        {
+            // Prepare the test data
+            string[] fields = new[] { "Id", "Name" };
+            IQueryable<Record> queryable = new SqliteQueryable(connection, "Course", "Alias", fields);
+
+            // Perform the test operation
+            Record record = queryable.SingleOrDefault();
+        }
+
+        [TestMethod]
+        public void SqlQueryable_SingleOrDefault_Predicate()
+        {
+            // Prepare the test data
+            string[] fields = new[] { "Id", "Name" };
+            IQueryable<Record> queryable = new SqliteQueryable(connection, "Course", "Alias", fields);
+
+            // Perform the test operation
+            Record record = queryable.SingleOrDefault(x => (int)x["Alias"]["Id"] == 2);
+
+            // Check the test result
+            Assert.AreEqual(2L, record["Alias"]["Id"]);
+        }
+    }
+}
