+++ src/LinqSql.Tests/Queryable/SqlQueryableLinqTests/SqlQueryableWhereTests.cs
@@ -1,0 +1,46 @@
+﻿using System.Data.Common;
+
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace System.Linq.Sql.Tests
+{
+    using Sqlite;
+
+    [TestClass]
+    public class SqlQueryableWhereTests
+    {
+        private readonly DbConnection connection = ConnectionTestHelper.CreatePopulatedConnection();
+
+        [TestMethod]
+        public void SqlQueryable_Where_Boolean()
+        {
+            // Prepare the test data
+            string[] fields = new[] { "Id", "Name" };
+            IQueryable<Record> queryable = new SqliteQueryable(connection, "Course", "Alias", fields);
+
+            // Perform the test operation
+            Record[] records = queryable
+                .Where(x => true)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(4, records.Length);
+        }
+
+        [TestMethod]
+        public void SqlQueryable_Where_Comparison()
+        {
+            // Prepare the test data
+            string[] fields = new[] { "Id", "Name" };
+            IQueryable<Record> queryable = new SqliteQueryable(connection, "Course", "Alias", fields);
+
+            // Perform the test operation
+            Record[] records = queryable
+                .Where(x => (int)x["Alias"]["Id"] == 1)
+                .ToArray();
+
+            // Check the test result
+            Assert.AreEqual(1, records.Length);
+        }
+    }
+}
